refactor(mirrormedia): add express-mini-apps/preview
