+++ packages/mirrormedia/keystone.ts
@@ -1,12 +1,12 @@
 import { config } from '@keystone-6/core'
 import { listDefinition as lists } from './lists'
 import appConfig from './config'
-import { createProxyMiddleware } from 'http-proxy-middleware'
 import envVar from './environment-variables'
 import express from 'express'
 import { createAuth } from '@keystone-6/auth'
 import { statelessSessions } from '@keystone-6/core/session'
 import { createApp as createPremiumMemberMiniApp } from './express-mini-apps/premium-membership/app'
+import { createPreviewMiniApp } from './express-mini-apps/preview/app'
 
 const { withAuth } = createAuth({
   listKey: 'User',
@@ -21,52 +21,6 @@ const { withAuth } = createAuth({
 })
 
 const session = statelessSessions(appConfig.session)
-
-function createPreviewMiniApp(createContext) {
-  const router = express.Router()
-
-  // Check if the request is sent by an authenticated user
-  const authenticationMw = async (req, res, next) => {
-    const context = await createContext(req, res)
-    // User has been logged in
-    if (context?.session?.data?.role) {
-      return next()
-    }
-
-    // Otherwise, redirect them to login page
-    res.redirect('/signin')
-  }
-
-  const previewProxyMiddleware = createProxyMiddleware({
-    target: envVar.previewServerOrigin,
-    changeOrigin: true,
-    onProxyRes: (proxyRes) => {
-      // The response from preview nuxt server might be with Cache-Control header.
-      // However, we don't want to get cached responses for `draft` posts.
-      // Therefore, we do not cache html response intentionlly by overwritting the Cache-Control header.
-      proxyRes.headers['cache-control'] = 'no-store'
-    },
-  })
-
-  // Proxy requests with `/story/id` url path to preview nuxt server
-  router.get('/story/:id', authenticationMw, previewProxyMiddleware)
-
-  // Proxy requests with `/event/:slug` url path to preview nuxt server
-  router.get('/event/:slug', authenticationMw, previewProxyMiddleware)
-
-  // Proxy requests with `/news/:id` url path to preview nuxt server
-  router.get('/news/:id', authenticationMw, previewProxyMiddleware)
-
-  // Proxy requests with `/_nuxt/*` url path to preview nuxt server
-  router.use(
-    '/_nuxt/*',
-    createProxyMiddleware({
-      target: envVar.previewServerOrigin,
-      changeOrigin: true,
-    })
-  )
-  return router
-}
 
 export default withAuth(
   config({
@@ -107,7 +61,12 @@ export default withAuth(
         app.use(jsonBodyParser)
 
         if (envVar.accessControlStrategy === 'cms') {
-          app.use(createPreviewMiniApp(createContext))
+          app.use(
+            createPreviewMiniApp({
+              previewServerOrigin: envVar.previewServerOrigin,
+              createContext,
+            })
+          )
         }
 
         if (envVar.accessControlStrategy === 'gql') {
